ci: fix CI build
- 4bfacc7 failed on all platforms

+++ checks/bugmalloc/test/data/ok.c
@@ -12,7 +12,7 @@
  */
 
 void*
-malloc(size_t sz);
+malloc(unsigned long long sz);
 #include "debugmalloc.h"
 
 void*
