+++ checks/bugmalloc/test/data/ok.cxx
@@ -11,7 +11,7 @@
  */
 
 extern "C" void*
-malloc(size_t sz);
+malloc(unsigned long long sz);
 
 #include "debugmalloc.h"
